[ADD] 머지 후 일부 조치

+++ src/main/java/com/likelion/mini2team/repository/DiaryRepository.java
@@ -4,5 +4,5 @@ import com.likelion.mini2team.domain.diary.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
-    void deleteById(Integer id);
+    void deleteById(long id);
 }
